Adjust how frames are shared between copies of animated visuals
