+++ src/Blitter.Bits/AnimatedVisual2D.cs
@@ -2,8 +2,10 @@ using System.Numerics;
 
 namespace Blitter.Bits;
 
-/// <summary>How an <see cref="AnimatedVisual2D"/> repeats once it
-/// runs past the last frame.</summary>
+/// <summary>
+/// How an <see cref="AnimatedVisual2D"/> repeats once it
+/// runs past the last frame.
+/// </summary>
 public enum AnimationLoop
 {
     /// <summary>Wrap back to the first frame.</summary>
@@ -15,16 +17,13 @@ public enum AnimationLoop
 }
 
 /// <summary>
-/// A <see cref="Visual2D"/> that cycles through frames of an
-/// <see cref="Atlas"/>. Playback is driven by the <c>elapsed</c>
-/// argument <see cref="Visual2D.Draw"/> receives, so a single
-/// instance can be shared across many hosts — each host's elapsed
-/// time is its own timeline.
+/// A <see cref="Visual2D"/> that cycles through frames of an <see cref="Atlas"/>. 
 /// </summary>
 public sealed class AnimatedVisual2D : Visual2D
 {
     private readonly Atlas _atlas;
-    private readonly int[] _frames;
+    // null = identity sequence (atlas frames 0..Count-1 in order); otherwise an explicit subset/reorder.
+    private readonly int[]? _frames;
     private BoundingCircle? _boundary;
 
     public AnimatedVisual2D(
@@ -39,16 +38,21 @@ public sealed class AnimatedVisual2D : Visual2D
             throw new ArgumentOutOfRangeException(nameof(frameDuration));
 
         _atlas = atlas;
-        _frames = frames.IsEmpty
-            ? BuildDefaultFrames(atlas.Count)
-            : frames.ToArray();
-        if (_frames.Length == 0)
-            throw new ArgumentException("At least one frame is required.", nameof(frames));
-        for (int i = 0; i < _frames.Length; i++)
+        if (frames.IsEmpty)
         {
-            if ((uint)_frames[i] >= (uint)atlas.Count)
-                throw new ArgumentOutOfRangeException(nameof(frames),
-                    $"Frame index {_frames[i]} is outside the atlas range [0, {atlas.Count}).");
+            if (atlas.Count == 0)
+                throw new ArgumentException("Atlas has no frames.", nameof(atlas));
+            _frames = null;
+        }
+        else
+        {
+            _frames = frames.ToArray();
+            for (int i = 0; i < _frames.Length; i++)
+            {
+                if ((uint)_frames[i] >= (uint)atlas.Count)
+                    throw new ArgumentOutOfRangeException(nameof(frames),
+                        $"Frame index {_frames[i]} is outside the atlas range [0, {atlas.Count}).");
+            }
         }
 
         FrameDuration = frameDuration;
@@ -56,47 +60,78 @@ public sealed class AnimatedVisual2D : Visual2D
         Offset = offset;
     }
 
-    /// <summary>Atlas providing the source regions.</summary>
+    // Clone ctor used by With* — skips validation and shares the (immutable-by-convention) frames array.
+    private AnimatedVisual2D(
+        Atlas atlas,
+        int[]? frames,
+        TimeSpan frameDuration,
+        AnimationLoop loop,
+        TimeSpan offset)
+    {
+        _atlas = atlas;
+        _frames = frames;
+        FrameDuration = frameDuration;
+        Loop = loop;
+        Offset = offset;
+    }
+
+    /// <summary>
+    /// Atlas providing the source regions.
+    /// </summary>
     public Atlas Atlas => _atlas;
 
-    /// <summary>Atlas region indices, in playback order.</summary>
-    public ReadOnlySpan<int> Frames => _frames;
-
-    /// <summary>Time each frame is held.</summary>
+    /// <summary>
+    /// Time each frame is held.
+    /// </summary>
     public TimeSpan FrameDuration { get; }
 
-    /// <summary>Loop behavior at the end of the sequence.</summary>
+    /// <summary>
+    /// Loop behavior at the end of the sequence.
+    /// </summary>
     public AnimationLoop Loop { get; }
 
-    /// <summary>Phase offset added to <c>elapsed</c> before picking a frame.
-    /// Use <see cref="WithOffset"/> to clone an instance out of phase.</summary>
+    /// <summary>
+    /// Phase offset added to <c>elapsed</c> before picking a frame.
+    /// Set at construction (or via <see cref="WithOffset"/>) to start a
+    /// new instance out of phase with another.
+    /// </summary>
     public TimeSpan Offset { get; }
 
-    /// <summary>Number of frames in the sequence.</summary>
-    public int FrameCount => _frames.Length;
+    /// <summary>
+    /// Number of frames in the sequence.
+    /// </summary>
+    public int FrameCount => _frames?.Length ?? _atlas.Count;
 
-    /// <summary>Returns a copy of this visual with a different phase
-    /// <paramref name="offset"/>; useful for desyncing hosts that
-    /// share an animation.</summary>
+    /// <summary>
+    /// Returns a copy of this visual with a different phase
+    /// <paramref name="offset"/>; useful for starting a second host's
+    /// animation out of phase.
+    /// </summary>
     public AnimatedVisual2D WithOffset(TimeSpan offset) =>
-        new(_atlas, FrameDuration, Loop, offset, _frames);
+        new(_atlas, _frames, FrameDuration, Loop, offset);
 
-    /// <summary>Returns a copy of this visual with a different
-    /// <paramref name="frameDuration"/>.</summary>
+    /// <summary>
+    /// Returns a copy of this visual with a different
+    /// <paramref name="frameDuration"/>.
+    /// </summary>
     public AnimatedVisual2D WithFrameDuration(TimeSpan frameDuration) =>
-        new(_atlas, frameDuration, Loop, Offset, _frames);
+        new(_atlas, _frames, frameDuration, Loop, Offset);
 
-    /// <summary>Returns a copy of this visual with a different
-    /// <paramref name="loop"/> behavior.</summary>
+    /// <summary>
+    /// Returns a copy of this visual with a different
+    /// <paramref name="loop"/> behavior.
+    /// </summary>
     public AnimatedVisual2D WithLoop(AnimationLoop loop) =>
-        new(_atlas, FrameDuration, loop, Offset, _frames);
+        new(_atlas, _frames, FrameDuration, loop, Offset);
 
-    /// <summary>Atlas region index that should be drawn for the given
-    /// elapsed time.</summary>
+    /// <summary>
+    /// Atlas region index that should be drawn for the given
+    /// elapsed time.
+    /// </summary>
     public int FrameIndexAt(TimeSpan elapsed)
     {
-        int n = _frames.Length;
-        if (n == 1) return _frames[0];
+        int n = FrameCount;
+        if (n == 1) return FrameAt(0);
 
         var step = (long)Math.Floor((elapsed + Offset).TotalSeconds / FrameDuration.TotalSeconds);
         long idx = Loop switch
@@ -106,7 +141,7 @@ public sealed class AnimatedVisual2D : Visual2D
             AnimationLoop.Once => step < 0 ? 0 : Math.Min(step, n - 1),
             _ => 0,
         };
-        return _frames[idx];
+        return FrameAt((int)idx);
 
         static long Mod(long a, int m)
         {
@@ -123,18 +158,23 @@ public sealed class AnimatedVisual2D : Visual2D
         }
     }
 
-    /// <summary>Atlas region for the current frame at the given time.</summary>
+    /// <summary>
+    /// Atlas region for the current frame at the given time.
+    /// </summary>
     public Rect RegionAt(TimeSpan elapsed) => _atlas[FrameIndexAt(elapsed)];
 
-    /// <summary>Bounding circle covering the largest frame in the sequence.</summary>
+    /// <summary>
+    /// Bounding circle covering the largest frame in the sequence.
+    /// </summary>
     public override BoundingCircle Boundary => _boundary ??= ComputeBoundary();
 
     private BoundingCircle ComputeBoundary()
     {
         float maxR2 = 0f;
-        for (int i = 0; i < _frames.Length; i++)
+        int n = FrameCount;
+        for (int i = 0; i < n; i++)
         {
-            var r = _atlas[_frames[i]];
+            var r = _atlas[FrameAt(i)];
             var hw = r.Width / 2f;
             var hh = r.Height / 2f;
             var r2 = hw * hw + hh * hh;
@@ -172,10 +212,5 @@ public sealed class AnimatedVisual2D : Visual2D
         }
     }
 
-    private static int[] BuildDefaultFrames(int count)
-    {
-        var a = new int[count];
-        for (int i = 0; i < count; i++) a[i] = i;
-        return a;
-    }
+    private int FrameAt(int i) => _frames is null ? i : _frames[i];
 }
